# Speedup to analysis.calculate_area