PR to fulfill #93. Substantial performance improvements from using scipy.ndimage.labeled_comprehension

Previous run-time:
`6.2 s ± 174 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`

New run-time:
`11.2 ms ± 194 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)`

I haven't added lat/lon area calculation, but this could also be done fairly easily