@w-k-jones 2022-03-28
I've added a new method to calculate area from 2D lat/lon coords. It uses a few approximations, so it's fairly inaccurate, but it'll do for now. I don't think it's worthwhile making any more updates to the analysis functions as they rely heavily on iris, and at some point we'll want to shift that over to xarray